+++ src/main/java/com/java/service/impl/VetServiceImpl.java
@@ -1,12 +1,14 @@
 package com.java.service.impl;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
+import com.java.exception.NotFoundException;
 import com.java.model.vet.Vet;
 import com.java.repository.VetRepo;
 import com.java.service.VetService;
@@ -28,7 +30,12 @@ public class VetServiceImpl implements VetService {
 	@Cacheable(cacheNames = "vet")
 	@Override
 	public Vet findVetById(Long id) {
-		return vetRepo.findById(id).get();
+		
+		Optional<Vet> vet = vetRepo.findById(id);
+		if(!vet.isPresent()) {
+			throw new NotFoundException("Vet with id : "+ id + " not found");
+		}
+		return vet.get() ;
 	}
 
 	
@@ -43,7 +50,7 @@ public class VetServiceImpl implements VetService {
 	public Vet updateVet(long id, Vet vet) {
 		Vet vetData = vetRepo.findById(id).orElse(null);
 		if(vetData== null) {
-			return null ;
+			throw new NotFoundException("Vet with id : "+ id + " not found");
 		}
 		vetData.setFirstName(vet.getFirstName());
 		vetData.setLastName(vet.getLastName());
@@ -56,6 +63,10 @@ public class VetServiceImpl implements VetService {
 	@CacheEvict(allEntries = true ,cacheNames ="vet")
 	@Override
 	public void deleteVet(Long id) {
+		Optional<Vet> vet = vetRepo.findById(id);
+		if(!vet.isPresent()) {
+			throw new NotFoundException("Vet with id : "+ id + " not found");
+		}
 		vetRepo.deleteById(id);
 	}
